api_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-cloud-storage==2.13.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
aiohttp==3.9.1
python-dateutil==2.8.2
gunicorn==21.2.0 